- add missing files\n - fix build script
