+++ src/main/java/org/svenson/info/methodhandle/MethodHandlePostConstructInvoker.java
@@ -1,4 +1,24 @@
 package org.svenson.info.methodhandle;
 
-public class MethodHandlePostConstructInvoker {
+import org.svenson.SvensonRuntimeException;
+import org.svenson.info.PostConstructInvoker;
+
+import java.lang.invoke.MethodHandle;
+
+class MethodHandlePostConstructInvoker implements PostConstructInvoker {
+
+    private final MethodHandle handle;
+
+    public MethodHandlePostConstructInvoker(MethodHandle handle) {
+        this.handle = handle;
+    }
+
+    @Override
+    public void invoke(Object o) {
+        try {
+            handle.invokeExact(o);
+        } catch (Throwable throwable) {
+            throw new SvensonRuntimeException(throwable);
+        }
+    }
 }
